new Provider and Recorder classes
