+++ src/Tizen.Sensor/Interop/Interop.Sensor.cs
@@ -131,104 +131,7 @@ internal static partial class Interop
 
         [DllImport(Libraries.Sensor, EntryPoint = "sensor_remove_sensor_removed_cb")]
         internal static extern int RemoveSensorRemovedCB(SensorRemovedCb callback);
-    }
-
-    internal static class SensorRecoder {
-        [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        internal delegate bool SensorRecorderDataCb(int type, int data, int remains, int error, Int64 userData);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_create_option")]
-        internal static extern int RecorderCreateOption( out IntPtr option);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_create_query")]
-        internal static extern int RecorderCreateQuery( out IntPtr query);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_data_get_double")]
-        internal static extern int RecorderDataGetDouble(int data, int key, out double value);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_data_get_time")]
-        internal static extern int RecorderDataGetTime(int data, out long start_time, out long end_time);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_destroy_option")]
-        internal static extern int RecorderDestroyOption(IntPtr option);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_destroy_query")]
-        internal static extern int RecorderDestroyQuery( IntPtr query);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_is_supported")]
-        internal static extern int RecorderIsSupported(int type, out bool isSupported);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_option_set_int")]
-        internal static extern int RecorderOptionSetInt(IntPtr option, int attribute , int value);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_query_set_int")]
-        internal static extern int RecorderQuerySetInt(IntPtr query, int attribute, int value);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_query_set_time")]
-        internal static extern int RecorderQuerySetTime(IntPtr query, int attribute, int time);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_read")]
-        internal static extern int RecorderRead(int type, IntPtr query, SensorRecorderDataCb cb , IntPtr user_data );
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_read_sync")]
-        internal static extern int RecorderReadSync(int type, IntPtr query, SensorRecorderDataCb cb, IntPtr user_data);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_start")]
-        internal static extern int RecorderStart(int type, IntPtr option);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_recorder_stop")]
-        internal static extern int RecorderStop(int type);
-
-    }
-
-    internal static class SensorProvider{
-        internal const int MAX_VALUE_SIZE = 16;
-
-        [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        internal delegate void SensorProviderStartCb(IntPtr provider, Int64 userData);
-
-        [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        internal delegate void SensorProviderStopCb(IntPtr provider, Int64 userData);
-
-        [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        internal delegate void SensorProviderIntervalChangedCb(IntPtr provider, uint IntervalMs, Int64 userData);
-   
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_create_provider")]
-        internal static extern int CreateProvider(String uri, out  IntPtr provider);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_add_provider")]
-        internal static extern int AddProvider(IntPtr provider);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_remove_provider")]
-        internal static extern int RemoveProvider(IntPtr provider);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_destroy_provider")]
-        internal static extern int DestroyProvider(IntPtr provider);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_set_name")]
-        internal static extern int ProviderSetName(IntPtr provider, String name);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_set_vendor")]
-        internal static extern int ProviderSetVendor(IntPtr provider, String vendor);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_set_range")]
-        internal static extern int ProviderSetRange(IntPtr provider, float minRange, float maxRange);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_set_resolution")]
-        internal static extern int ProviderSetResolution(IntPtr provider, float resolution);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_set_start_cb")]
-        internal static extern int SetProviderStartCb(IntPtr provider, SensorProviderStartCb startCb, IntPtr userData);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_set_stop_cb")]
-        internal static extern int SetProviderStopCb(IntPtr provider, SensorProviderStopCb stopCb, IntPtr userData);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_set_interval_changed_cb")]
-        internal static extern int SetProviderIntervalChangedCb(IntPtr provider, SensorProviderIntervalChangedCb intervalCb, IntPtr userData);
-
-        [DllImport(Libraries.Sensor, EntryPoint = "sensor_provider_publish")]
-        internal static extern int ProviderPublish(IntPtr provider, Tizen.Sensor.SensorEventStruct _event);
-    }
+    } 
 
     internal static partial class Libc
     {
@@ -263,4 +166,14 @@ internal static partial class Interop
         SensorStressMonitorData outStruct = (SensorStressMonitorData)Marshal.PtrToStructure<SensorStressMonitorData>(unmanagedVariable);
         return outStruct;
     }
+    internal static SensorEventStruct ClassToEventStruct(SensorEvent obj)
+    {
+        SensorEventStruct outStruct = new SensorEventStruct();
+        outStruct.accuracy = obj.Accuracy;
+        outStruct.timestamp = obj.Timestamp;
+        outStruct.value_count = obj.ValueCount;
+        outStruct.values = obj.GetValues();
+        return outStruct;
+    }
+
 }
