+++ src/Tizen.Sensor/Tizen.Sensor/EventArgs/StressMonitorDataUpdatedEventArgs.cs
@@ -23,7 +23,7 @@ namespace Tizen.Sensor
     /// Values of the stress monitor sensor
     /// </summary>
     [StructLayout(LayoutKind.Sequential)]
-    public struct StressMonitorValue : IEquatable<StressMonitorValue>
+    public class StressMonitorValue
     {
         /// <summary> 
         /// Time stamp value 
@@ -69,70 +69,13 @@ namespace Tizen.Sensor
             get;
             private set;
         }
-
-        /// <summary>
-        /// Gets the hash code for this instance of <see cref="StressMonitorValue"/>.
-        /// </summary>
-        /// <returns>The hash code for this instance of <see cref="StressMonitorValue"/>.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public override int GetHashCode()
-        {
-            return new { TimeStamp, Flag, Stress, Progress, InfoAlgo }.GetHashCode();
-        }
-
-        /// <summary>
-        /// Compares two instances of <see cref="StressMonitorValue"/> for inequality.
-        /// </summary>
-        /// <param name="monitorValue1">A <see cref="StressMonitorValue"/> to compare.</param>
-        /// <param name="monitorValue2">A <see cref="StressMonitorValue"/> to compare.</param>
-        /// <returns>true if the two instances of <see cref="StressMonitorValue"/> are equal; otherwise false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public static bool operator !=(StressMonitorValue monitorValue1, StressMonitorValue monitorValue2)
-        {
-            return !(monitorValue1 == monitorValue2);
-        }
-
-        /// <summary>
-        /// Compares two instances of <see cref="StressMonitorValue"/> for inequality.
-        /// </summary>
-        /// <param name="monitorValue1">A <see cref="StressMonitorValue"/> to compare.</param>
-        /// <param name="monitorValue2">A <see cref="StressMonitorValue"/> to compare.</param>
-        /// <returns>true if the two instances of <see cref="StressMonitorValue"/> are not equal; otherwise false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public static bool operator ==(StressMonitorValue monitorValue1, StressMonitorValue monitorValue2)
-        {
-            return !(monitorValue1 != monitorValue2);
-        }
-
-        /// <summary>
-        /// Compares an object to an instance of <see cref="StressMonitorValue"/> for equality.
-        /// </summary>
-        /// <param name="obj">A <see cref="Object"/> to compare.</param>
-        /// <returns>true if the points are equal; otherwise, false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public override bool Equals(object obj)
-        {
-            return obj is StressMonitorValue && this == (StressMonitorValue)obj;
-        }
-
-        /// <summary>
-        /// Compares an object to an instance of <see cref="StressMonitorValue"/> for equality.
-        /// </summary>
-        /// <param name="other">A <see cref="Object"/> to compare.</param>
-        /// <returns>true if the points are equal; otherwise, false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public bool Equals(StressMonitorValue other)
-        {
-            return this == (StressMonitorValue)other;
-        }
-
     }
 
     /// <summary>
     /// Data of the stress monitor
     /// </summary>
     [StructLayout(LayoutKind.Sequential)]
-    public struct SensorStressMonitorData : IEquatable<SensorStressMonitorData>
+    public class SensorStressMonitorData
     {
         internal const int SENSOR_STRESS_MONITOR_DATA_SIZE = 600;
         internal const int SENSOR_STRESS_MONITOR_BUF_SIZE = 44;
@@ -171,10 +114,9 @@ namespace Tizen.Sensor
         /// <summary> 
         /// Array of StressMonitorValue
         /// </summary>
-        public StressMonitorValue[] Values
+        public StressMonitorValue[] GetValues()
         {
-            get;
-            private set;
+            return values;
         }
 
         /// <summary>
@@ -193,66 +135,9 @@ namespace Tizen.Sensor
         /// <summary> 
         /// Array of Histogram values
         /// </summary>
-        public UInt32[] Histogram
+        public UInt32[] GetHistogram()
         {
-            get { return histogram; }
-            private set { }
-        }
-
-        /// <summary>
-        /// Gets the hash code for this instance of <see cref="SensorStressMonitorData"/>.
-        /// </summary>
-        /// <returns>The hash code for this instance of <see cref="SensorStressMonitorData"/>.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public override int GetHashCode()
-        {
-            return new { Mode, Accuracy, ValueCount, Values, BaseHr, Histogram }.GetHashCode();
-        }
-
-        /// <summary>
-        /// Compares two instances of <see cref="SensorStressMonitorData"/> for inequality.
-        /// </summary>
-        /// <param name="monitorData1">A <see cref="SensorStressMonitorData"/> to compare.</param>
-        /// <param name="monitorData2">A <see cref="SensorStressMonitorData"/> to compare.</param>
-        /// <returns>true if the two instances of <see cref="SensorStressMonitorData"/> are equal; otherwise false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public static bool operator !=(SensorStressMonitorData monitorData1, SensorStressMonitorData monitorData2)
-        {
-            return !(monitorData1 == monitorData2);
-        }
-
-        /// <summary>
-        /// Compares two instances of <see cref="SensorStressMonitorData"/> for inequality.
-        /// </summary>
-        /// <param name="monitorData1">A <see cref="SensorStressMonitorData"/> to compare.</param>
-        /// <param name="monitorData2">A <see cref="SensorStressMonitorData"/> to compare.</param>
-        /// <returns>true if the two instances of <see cref="SensorStressMonitorData"/> are not equal; otherwise false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public static bool operator ==(SensorStressMonitorData monitorData1, SensorStressMonitorData monitorData2)
-        {
-            return !(monitorData1 != monitorData2);
-        }
-
-        /// <summary>
-        /// Compares an object to an instance of <see cref="SensorStressMonitorData"/> for equality.
-        /// </summary>
-        /// <param name="obj">A <see cref="Object"/> to compare.</param>
-        /// <returns>true if the points are equal; otherwise, false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public override bool Equals(object obj)
-        {
-            return obj is SensorStressMonitorData && this == (SensorStressMonitorData)obj;
-        }
-
-        /// <summary>
-        /// Compares an object to an instance of <see cref="SensorStressMonitorData"/> for equality.
-        /// </summary>
-        /// <param name="other">A <see cref="Object"/> to compare.</param>
-        /// <returns>true if the points are equal; otherwise, false.</returns>
-        /// <since_tizen> 6 </since_tizen>
-        public bool Equals(SensorStressMonitorData other)
-        {
-            return this == (SensorStressMonitorData)other;
+            return histogram;
         }
     }
